+++ packages/admin-ui/src/page/header.tsx
@@ -1,7 +1,7 @@
 /**
  * WordPress dependencies
  */
-import { Stack } from '@wordpress/ui';
+import { Stack, Text } from '@wordpress/ui';
 
 /**
  * Internal dependencies
@@ -52,9 +52,13 @@ export default function Header( {
 						</div>
 					) }
 					{ title && (
-						<HeadingTag className={ styles[ 'header-title' ] }>
+						<Text
+							className={ styles[ 'header-title' ] }
+							render={ <HeadingTag /> }
+							variant="heading-lg"
+						>
 							{ title }
-						</HeadingTag>
+						</Text>
 					) }
 					{ breadcrumbs }
 					{ badges }
@@ -70,7 +74,13 @@ export default function Header( {
 				</Stack>
 			</Stack>
 			{ subTitle && (
-				<p className={ styles[ 'header-subtitle' ] }>{ subTitle }</p>
+				<Text
+					render={ <p /> }
+					variant="body-md"
+					className={ styles[ 'header-subtitle' ] }
+				>
+					{ subTitle }
+				</Text>
 			) }
 		</Stack>
 	);